Add a "clue" plugin
This will be used for games like "Pyramid" and "Tri, Tri Again" where
it will be handy to have a bot keeping the channel apprised of the
current clue.

To start with, this plugin is simply repeating things told to it via
direct message, (this is literally just a copy of the repeat.py plugin
from the docs directory).