Add `jits guess` functionality
Players can now type `jits guess [word]` to add it to a list of
outstanding guesses.

If the word is already a jewel, sand, or an outstanding guess, stony
will let the players know.

A guess is removed from the outstanding guess list when the Sultan
categorizes it as a jewel or sand.