Add some convenience commands (aliases) for guesses and ruling on guesses
These new aliases do not follow the "jit <command> <args...>" format,
but instead look like this:

	fish?
	fish=jewel

	fowl?
	fowl=sand

That is, it's the exact syntax originally used by players of the game
before we had the bot involved.

The mechanics of the alias implementation is that there is a table of
regular-expression substutions to turn an alias into the equivalent
command, (such that "fish?" would become "jits guess fish", etc.), and
then the conventional command processing follows.