Fix bug preventing list from being shown when Sultan added new words
Since the addition of support for tracking guesses, there's been a bug
in stony, as follows: If the Sultan ruled on a previously-guessed
word, everything would work. But, if the Sultan was feeling
magnanimous and ruled on a word that had not previously been guessed,
it would still be added to the correct list, but stony would stay
silent about it, (not report the ruling, nor the updated list).

The bug came about due to the line of code that called:

	guess_list.remove(word)

In the case where guess_list does not actually contain the word, the
above code results in a ValueError exception. Now, we also have a bug
in the higher-level rtmbot code that is causing our bot to silently
ignore exceptions like this in our plugins, (the stack trace should at
least be printed on stderr or so). So that makes simple bugs like this
harder to find/fix than they should be.

Anyway, in this commit we check that the word is actually in the list
before calling the remove() method on the list.

[Note: There's a part of me that cringes to be calling "in" before
"remove" since that's necessarily twice as much work. The more
''efficient'' code might look something like:

	try:
	    guess_list.remove(word)
	except ValueError:
	    pass

but that code is so ugly that it would make me cringe even more.]